The create and update rings and segmetns have been refactored
