+++ src/main/java/com/a5lab/axion/domain/wizard/RadarProcessor.java
@@ -1,4 +1,4 @@
-package com.a5lab.axion.domain.wizard.processors;
+package com.a5lab.axion.domain.wizard;
 
 public interface RadarProcessor {
   void process() throws Exception;
