+++ src/main/java/com/a5lab/axion/domain/wizard/WizardServiceImpl.java
@@ -6,9 +6,6 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.a5lab.axion.domain.wizard.processors.RadarProcessor;
-import com.a5lab.axion.domain.wizard.processors.RadarProcessorFactory;
-
 @RequiredArgsConstructor
 @Service
 @Transactional
